Add command to assign inductees role based on CSV
The new /assigninductees application command reads from a local CSV
file, extracts the Discord usernames, and gives those members the
@Inductees roles.

The implementation tries its best in error handling. It tries to assign
the role to as many members as possible (instead of short-circuiting on
the first error). Members that aren't found or those for which a Discord
API error occurred are collected and outputted as part of the response.