Report mentions of members actually updated
This entailed a pretty hefty refactoring of the logic layer. The making
and error-handling of the actual API calls were refactored into their
own functions, and they communicate result status with trivalent logic
instead of a boolean:

* SUCCESS: a request was made and it succeeded.
* FAILED: a request was made and it failed.
* SKIPPED: a request was not made because there was no need to.

The presentation layer (prepareResponseEmbed) was also cleaned up to use
helper closure functions for each part of the body.