Add /linktree as shortcut for induction Linktree
This does NOT hard-code the Linktree texts and hyperlinks! This fetches
the HTML of the actual UPE induction Linktree and parses out the
categories and link buttons. They are then formatted into hyperlinks
within an embed and displayed to the caller.

This addBroadcastOption helper was added to de-duplicate the pattern of
adding a "broadcast" Boolean option to commands (specifying a public
response instead of a default ephemeral response).