feat!: save img filename, not Discord URL, to DB
This makes it so that the event.pic field uses the filename of the image
downloaded from Discord to the local event-pics/ directory. The event
pages rendered from /viewevents then uploads these local files as
attachments for the embed images. However, note that the performance
becomes very laggy this way, presumably because that images are
"attached" over and over again as the user changes pages with the arrow
buttons.

BREAKING CHANGE: the event.pic fields for all documents in the bytes
collection need to be changed to be the filename of the image, not the
Discord attachment URL. The bot crashes otherwise as it would
incorrectly try to treat a URL as a file path.