feat!: disable bit-byte leaderboard/events cmds (#25)
## Motivation

As of F25 induction season, the bit-byte program no longer has a
leaderboard/prize system.

## Summary

1. Added a new property to the `SlashCommandHandler` ABC defining
whether the handler definition should be deployed (and thus be invocable
on Discord). This defaults to `true`; setting to `false` is to express
that we might still want this definition around for whatever reason but
just don't deploy it for now.
2. Deprecate and disable command registration for the
`/leaderboardbitbyte`, `/submitbitbyte`, `/jeopardybitbyte`, and
`/eventsbitbyte` commands.