chore: update bit-byte features for F25 (#26)
## Motivation

Bringing the leaderboard & bit-byte events features back from the grave
(#25) by popular request, and also cleaning it up a bit to align with
new bit-byte details. For context, as of F25 induction season:

- The bit-byte program is no longer tied to the induction process
itself. It has been opened up to the general community, not just to
inductees.
- There is no longer a prize system, but bytes expressed interest in
continuing to have a leaderboard for competitive spirit.

## Summary

Re-enabled the commands disabled in #25 but also:

- `/leaderboardbitbyte` now handles ties in the display.
- `/eventsbitbyte` no longer spawns a group selector component to then
spawn the event picture embeds. The command now takes in a role as a
required option and directly replies with the paginated event picture
events. I really didn't like the former flow. The current bot framework
also does not work well with component interactions & dispatching in
general, so I would rather avoid it where I can.
- The bit-byte score calculation no longer depends on the total number
of bits in a family. This is because it is hard to get this number.
Before, we had the invariant that every inductee was in the server, so
we could just count || `@Inductees` $\cap$ `@Family-Role` ||. Now, the
program has been opened up to beyond inductees who may or may not join
the server (sigh). We could somehow make the bot aware of actual
bit-byte pairings by way of seasonal hard-coding or spreadsheet or
config file but that's a lot of trouble and further complicates
maintenance.
- Misc cleanup. Look at commit history attached to this PR.