chore: add channel checking fn to handle threads
This serves as a shortcut for checking if a certain channel object OR
its parent channel satisfy some condition. This is useful for e.g.
checking if a message was sent in either a channel directly or threads
within that channel.

This commit also refactors the arf listener to use this helper in
calculation of whether the channel is a bit-byte channel. It also now
covers the previously neglected case of threads under an induction
channel.